requirements.txt
flask==2.3.3
pandas==2.1.1
numpy==1.24.3
yfinance==0.2.22
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.21
flask-cors==4.0.0
websockets==11.0.3
pytest==7.4.2
backtrader==1.9.78.123
ta-lib==0.4.28
matplotlib==3.7.2
plotly==5.17.0